Experimental implementation of callback, which is inefficient
I could not use the reference in the argument of callback and make its
lifetime work with the tape itself, so the nodes are copied as a vector
and passed to the callback, which is extremely inefficient.
It would be rather more efficient to borrow-unborrow in every access to
TapeNode.