Don't evaluate the same value twice in `eval()`
Previously, some nodes are visited more than once to evaluate the
expression, if a node has a common ancestor. The most obvious example is
a square, x * x. However, unlike backprop, we don't need to visit them
multiple times, because the result is always the same (unless you put
randomness to unary function, which is bad).