Add UnaryFn support
We use Option to allow taking away ownership of the field temporarily in
order to avoid invalidating the nodes buffer while potentially
generating a node in gen_graph. I'm not sure if it's the best way to do
it. Another way would be to use Rc, but the reference counting cost may
be greater than swapping a pointer only in gen_graph, which should not
happen often.